% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_arEstBipTau.R
\name{ar_est_bip_tau}
\alias{ar_est_bip_tau}
\title{ar_est_bip_tau}
\usage{
ar_est_bip_tau(x, P)
}
\arguments{
\item{x:}{data (observations/measurements/signal)}

\item{P:}{autoregressive order}
}
\value{
phi_hat: a vector of bip tau estimates for each order up to P

x_filt: cleaned version of x using robust BIP predictions

a_scale_final: minimal tau scale of the innovations of BIP AR or AR
}
\description{
The function ar_est_bip_tau(x,P) comuptes BIP S-estimates of the
AR model parameters. It also computes an outlier cleaned signal using
BIP-AR(P) predictions, and the M-scale of the estimated innovations
series.
}
\examples{

}
\references{
"Robust Statistics for Signal Processing"
Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
Cambridge University Press, 2018.

"Bounded Influence Propagation \eqn{\tau}-Estimation: A New Robust Method for ARMA Model Estimation."
Muma, M. and Zoubir, A.M.
IEEE Transactions on Signal Processing, 65(7), 1712-1727, 2017.
}
