% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_armaEstBipMM.R
\name{arma_est_bip_mm}
\alias{arma_est_bip_mm}
\title{arma_est_bip_mm}
\usage{
arma_est_bip_mm(x, p, q, tolX = 5e-07)
}
\arguments{
\item{x:}{data (observations/measurements/signal)}

\item{p:}{autoregressive order}

\item{q:}{moving-average order}
}
\value{
result: named list with following fields
                \item{ar_coeffs}{numeric vector of length p. BIP-AR(p) MM estimates}
                \item{ma_coeffs}{numeric vector of length q. BIP-AR(q) MM estimates}
                \item{inno_scale}{numeric, BIP s-estimate of the innovations scale}
                \item{ar_coeffs_init}{numeric vector of length p. Robust starting point for estimation}
                \item{ma_coeffs_init}{numeric vector of length q. Robust starting point for estimation}
}
\description{
The function  arma_est_bip_mm(x,p,q) comuptes BIP MM-estimates of the
ARMA model parameters.
It also computes an outlier cleaned signal using BIP-ARMA(p,q) predictions
}
\note{
file is in dependentData_armaEstBipMM.R
}
\examples{
library(signal)

N <- 500
a <- rnorm(N)
p <- 1
q <- 0
x <- signal::filter(1, c(1, -0.8), a)

arma_est_bip_mm(x, p, q)
}
\references{
"Robust Statistics for Signal Processing"
  Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
  Cambridge University Press, 2018.

 "Bounded Influence Propagation \eqn{\tau}-Estimation: A New Robust Method for ARMA Model Estimation."
  Muma, M. and Zoubir, A.M.
  IEEE Transactions on Signal Processing, 65(7), 1712-1727, 2017.
}
