% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary_armaResid.R
\name{arma_s_resid}
\alias{arma_s_resid}
\title{arma_s_resid}
\usage{
arma_s_resid(x, beta_hat, p, q)
}
\arguments{
\item{x:}{numeric vector. The signal}

\item{beta_hat:}{the AR and MA coefficients}

\item{p:}{AR order}

\item{q:}{MA order}
}
\value{
a: numeric vector of residuals
}
\description{
Computes the residuals of S-estimates for an ARMA process
}
\note{
File is in dependentData_Auxiliary_armaResid.R
}
