% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependentData_Auxiliary_arma_s_resid_sc.R
\name{arma_s_resid_sc}
\alias{arma_s_resid_sc}
\title{arma_s_resid_sc}
\usage{
arma_s_resid_sc(x, beta_hat, p, q)
}
\arguments{
\item{x}{: numeric vector. The signal}

\item{beta_hat}{: Initial estimate of ARMA coefficients}

\item{p}{: AR order}

\item{q}{: MA order}
}
\value{
a_sc : scale estimate
}
\description{
arma_s_resid_sc
}
\note{
File location : DependentData_Auxiliary_arma_s_resid_sc.R
}
\examples{
data("x_ao_arma")
arma_s_resid_sc(x_ao_arma, c(-.8, .5), 1, 1)

library(signal)
x <- rnorm(10)
sig <- signal::filter(-.8, 1, x)
arma_s_resid_sc(sig, c(-.8), 1, 0)
}
