% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustFiltering_asymmetricTanh.R
\name{asymmetric_tanh}
\alias{asymmetric_tanh}
\title{asymmetric_tanh}
\usage{
asymmetric_tanh(sig, c1, c2, x1)
}
\arguments{
\item{sig}{: the signal}

\item{c1}{: first clipping point}

\item{c2}{: second clipping point}

\item{x1}{: smoothing parameter to make the score function continuous}
}
\value{
phi : asymmetric tanh score function

phi_point : derivative of asymmetric score function
}
\description{
Computes the smoothed asymmetric tanh score
and its derivative.
}
\examples{

asymmetric_tanh(c(1, 3, 4, 5, 62, 346, 23562, 65, 264, 21362, 90), 6, 87, 4)
}
\references{
"Robust Statistics for Signal Processing"
 Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
 Cambridge University Press, 2018.

 "Robust Tracking and Geolocation for Wireless Networks in NLOS Environments."
  Hammes, U., Wolsztynski, E., and Zoubir, A.M.
  IEEE Journal on Selected Topics in Signal Processing, 3(5), 889-901, 2009.
}
