% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary_armaResid.R
\name{bip_s_resid_sc}
\alias{bip_s_resid_sc}
\title{bip_s_resid_sc}
\usage{
bip_s_resid_sc(x, beta_hat, p, q)
}
\description{
bip_s_resid_sc
}
\note{
file in dependentData_Auxiliary_armaResid.R
}
