% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_spec_arma_est_bip.R
\name{bip_tau_arma_PSD_book}
\alias{bip_tau_arma_PSD_book}
\title{bip_tau_arma_PSD_book}
\usage{
bip_tau_arma_PSD_book(x, p, q)
}
\description{
bip_tau_arma_PSD_book
}
\note{
File located in SpectrumEstimation_spec_arma_est_bip.R
}
