% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary_armaResid.R
\name{bip_tau_resid_sc}
\alias{bip_tau_resid_sc}
\title{bip_tau_resid_sc}
\usage{
bip_tau_resid_sc(x, beta_hat, p, q)
}
\description{
bip_tau_resid_sc
}
\note{
In file dependentData_Auxiliary_armaResid
}
\examples{
data("x_ao_mat")

# 4.2918
res <- tau_scale(x_ao_mat)

## Example 2

data("PPG_Data")

library(zeallot)

ibi_ppg <- diff(PPG_Data$ppg.pos[1,])

x <- ibi_ppg[length(ibi_ppg):1] - median(ibi_ppg)

beta_mat <- c(-0.6640 ,  -0.0815 ,  -0.2180 ,  -0.4591 ,  -0.0830 ,   0.2936 ,  -0.0434   , 0.0190 ,   0.0757 ,  -0.0938,   -0.1758)

bip_tau_resid_sc(x, beta_mat, 0, 11)

}
