% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_biweight_filter.R
\name{biweight_filter}
\alias{biweight_filter}
\title{biweight_filter}
\usage{
biweight_filter(x)
}
\arguments{
\item{x:}{data (observations/measurements/signal), real-valued vector}
}
\value{
xFBi: Biweight filtered (outlier cleaned) signal

ABi:  Fourier coefficients for cosine

BBi:  Fourier coefficients for sine
}
\description{
The biweight_filter(x) is our implementation of the
method described in
}
\details{
"High breakdown methods of time series analysis.
Tatum, L.G., and Hurvich, C. M.
Journal of the Royal Statistical Society. Series B (Methodological),
pp. 881-896, 1993.

The code is based on an implementation by Falco Strasser, Signal Processing
Group, TU Darmstadt, October 2010.
}
\note{
File location: SpectrumEstimation_biweight_filter.R
}
\examples{

}
\references{
"Robust Statistics for Signal Processing"
Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
Cambridge University Press, 2018.
}
