% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustFiltering_createEnvironmentBook.R
\name{create_environment_book}
\alias{create_environment_book}
\title{create_environment_book}
\usage{
create_environment_book(parameter, start, sigma_v)
}
\description{
The function computes generates data for tracking of a mobile user
equipment based on the values received by parameter. It is based on a function written by Ulrich Hammes,
Signal Processing Group, TU Darmstadt, February 2009
}
