% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustFiltering_ekfMasreliez.R
\name{ekf_toa_Masreliez}
\alias{ekf_toa_Masreliez}
\title{ekf_Masreliez}
\usage{
ekf_toa_Masreliez(r_ges, theta_init, BS, parameter = NULL)
}
\description{
Masreliez type EKF for tracking with time-of-arrival (ToA) estimates.
}
\note{
requires following libraries:
   * pracma
   * zeallot
}
