% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustFiltering_ekfToARobust.R
\name{ekf_toa_robust}
\alias{ekf_toa_robust}
\title{ekf_toa}
\usage{
ekf_toa_robust(r_ges, theta_init, BS, parameter = NULL)
}
\arguments{
\item{r_ges}{: measured distances as a M x N matrix}

\item{theta_init}{: initial state estimate}

\item{BS}{: base station positions}

\item{parameter}{:}
}
\value{
th_hat : state estimates

P_min : apriori covariance

P : aposteriors covariance
}
\description{
Robust EKF for tracking with time-of-arrival (:= ToA) estimates
based on M-estimation.
}
\examples{
ekf_toa(matrix(rnorm(16), 4, 4),
        rnorm(4),
        matrix(rnorm(8), 4, 2))
}
\references{
"Robust Statistics for Signal Processing"
 Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
 Cambridge University Press, 2018.

 "Robust Tracking and Geolocation for Wireless Networks in NLOS Environments."
  Hammes, U., Wolsztynski, E., and Zoubir, A.M.
  IEEE Journal on Selected Topics in Signal Processing, 3(5), 889-901, 2009.
}
