% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary.R
\name{eta}
\alias{eta}
\title{eta}
\usage{
eta(x, c = 1)
}
\arguments{
\item{x:}{numeric vector, the signal}

\item{c:}{numeric, default = 1}
}
\description{
eta
}
\note{
file location : dependentData_Auxilliary.R
}
\examples{
library(Rrobustsp)

x <- rnorm(5)
eta(x)

}
