% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{hubreg}
\alias{hubreg}
\title{hubreg computes the joint M-estimates of regression and scale using
  Huber's criterion. Function works for both real- and complex-valued data.}
\usage{
hubreg(y, X, c = NULL, sig0 = NULL, b0 = NULL, printitn = 0,
  iter_max = 2000, errortol = 1e-05)
}
\arguments{
\item{y:}{Numeric data vector of size N x 1 (output, respones)}

\item{X:}{Numeric data matrix of size N x p. Each row represents one
observation, and each column represents one predictor (feature).
If the model has an intercept, then first column needs to be a
vector of ones.}

\item{c:}{numeric threshold constant of Huber's function}

\item{sig0:}{(numeric) initial estimator of scale [default: SQRT(1/(n-p)*RSS)]}

\item{b0:}{initial estimator of regression (default: LSE)}

\item{printitn:}{print iteration number (default = 0, no printing)}

\item{iter_max:}{maximum number of iterations. \cr default = 2000}

\item{errortol:}{ERROR TOLERANCE FOR HALTING CRITERION. \cr default = 1e-5}
}
\value{
b1: the regression coefficient vector estimate

sig1: the estimate of scale

iter: the # of iterations
}
\description{
hubreg computes the joint M-estimates of regression and scale using
  Huber's criterion. Function works for both real- and complex-valued data.
}
