% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_ladlasso.R
\name{ladlasso}
\alias{ladlasso}
\title{ladlasso}
\usage{
ladlasso(y, X, lambda, intcpt = T, b0 = NULL, reltol = 1e-08,
  printitn = 0, iter_max = 2000)
}
\arguments{
\item{y}{numeric response N x 1 vector (real/complex)}

\item{X}{sparse matrix, numeric feature  N x p matrix (real/complex)}

\item{lambda}{numeric, non-negative penalty parameter}

\item{intcpt}{numeric optional initial start of the regression vector for
IRWLS algorithm. If not given, we use LSE (when p>1).}

\item{b0}{numeric optional initial start of the regression vector for
IRWLS algorithm. If not given, we use LSE (when p>1).}

\item{reltol}{Convergence threshold for IRWLS. Terminate when successive
estimates differ in L2 norm by a rel. amount less than reltol.}

\item{printitn}{print iteration number (default = 0, no printing)}

\item{iter_max}{number of iterations \cr
default = 2000}
}
\value{
b1 numeric, the regression coefficient vector of size N

integr, iter number of iterations
}
\description{
ladlasso computes the LAD-Lasso regression estimates for given complex-
or real-valued data.  If number of predictors, p, is larger than one,
then IRWLS algorithm is used, otherwise a weighted median algorithm
(N > 200) or elemental fits (N<200).
}
