% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{ladreg}
\alias{ladreg}
\title{ladrag}
\usage{
ladreg(y, X, intcpt = T, b0 = NULL, printitn = 0)
}
\arguments{
\item{y:}{numeric response N x 1 vector (real/complex)}

\item{X:}{numeric feature  N x p matrix (real/complex)}

\item{intcpt:}{(logical) flag to indicate if intercept is in the model}

\item{b0:}{numeric optional initial start of the regression vector for
IRWLS algorithm. If not given, we use LSE (when p>1).}

\item{printitn:}{print iteration number (default = 0, no printing) and
other details}
}
\value{
b1: (numberic) the regression coefficient vector

iter: (numeric) # of iterations (given when IRWLS algorithm is used)
}
\description{
ladreg computes the LAD regression estimate
}
\examples{

}
