% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{madn}
\alias{madn}
\title{madn computes the normalized median absolute deviation estimate of
scale, i.e.,}
\usage{
madn(y)
}
\description{
mu_hat = arg min_mu SUM_i rho_TUK(y_i - mu)
}
\details{
INPUTS:
          y: data vector of size N x 1
  OUTPUT:
          sig: normalized median absolute deviations scale estimate
}
