% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mat_stem}
\alias{mat_stem}
\title{mat_stem}
\usage{
mat_stem(x, y, pch = 16, linecol = 1, clinecol = 1, ...)
}
\description{
matlab-like stem plot
}
\references{
https://www.r-bloggers.com/matlab-style-stem-plot-with-r/
}
