% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location_Scale.R
\name{Mlocscale}
\alias{Mlocscale}
\title{Mlocscale computes Huber's M-estimates of location and scale.}
\usage{
Mlocscale(y, c = NULL)
}
\arguments{
\item{y:}{real valued data vector of size N x 1}

\item{c:}{tuning constant c>=0}
}
\value{
mu_hat: Huber's M-estimate of location

sigma_hat: Huber's M-estimate of scale
}
\description{
Mlocscale computes Huber's M-estimates of location and scale.
}
