% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{Mreg}
\alias{Mreg}
\title{Mreg computes the M-estimates of regression using an auxiliary scale
estimate. It uses the iterative reweighted least squares (IRWLS) algorithm}
\usage{
Mreg(y, X, lossfun = "huber", b0 = NULL, verbose = F)
}
\arguments{
\item{y}{: (numeric) data vector of size N (output, response vector)}

\item{X}{: (numeric) data matrix of size N x p (input, feature matrix)
If the model has intercept, then first column of X should be a
vector of ones.}

\item{lossfun}{: (string) either 'huber' or 'tukey' to identify the desired
loss function. Default is 'huber'}

\item{b0}{: (numeric) Optional robust initial start (regression vector) of
iterations. If not given, we use the LAD regression estimate}

\item{verbose:}{(logical) true of false (default). Set as true if you wish
to see convergence as iterations evolve.}
}
\value{
b1 : regression parameters

sig: scale
}
\description{
Mreg computes the M-estimates of regression using an auxiliary scale
estimate. It uses the iterative reweighted least squares (IRWLS) algorithm
}
\examples{
Mreg(1:5, matrix(-1:3))
Mreg(1:5, matrix(-1:3), lossfun = 'tukey')
}
