% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location_Scale.R
\name{MscaleTUK}
\alias{MscaleTUK}
\title{% Mscale_TUK computes Tukey's M-estimate of
% scale.
%
%
%   INPUTS:
%           y: real valued data vector of size N x 1
%           c: tuning constant c>=0
%
%   OUTPUT:
%           sigma_hat: Tukey's M-estimate of scale}
\usage{
MscaleTUK(y, c = 4.685, max_iters = 1000, tol_err = 1e-05)
}
\description{
% Mscale_TUK computes Tukey's M-estimate of
% scale.
%
%
%   INPUTS:
%           y: real valued data vector of size N x 1
%           c: tuning constant c>=0
%
%   OUTPUT:
%           sigma_hat: Tukey's M-estimate of scale
}
