% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_Mscat.R
\name{Mscat}
\alias{Mscat}
\title{Mscat}
\usage{
Mscat(x, loss, losspar = NULL, invC = NULL, printitn = 0)
}
\arguments{
\item{x}{: n x p matrix}

\item{loss}{: 'Huber', 't_loss' or 'Tyler'}

\item{losspar:}{parameter of the loss function: q in [0,1) for Huber and
d.o.f. v >= 0 for t-loss. For Tyler you do not need to specify
this value. Parameter q determines the treshold
c^2 as the qth quantile of chi-squared distribution with p
degrees of freedom distribution (Default q = 0.8). Parameter v
is the def.freedom of t-distribution (Default v = 3)
if v = 0, then one computes Tyler's M-estimator}

\item{invC:}{initial estimate is the inverse scatter matrix (default =
inverse of the sample covariance matrix)}

\item{printitn:}{integer, print iteration number (default = 0, no printing)}
}
\value{
C: the M-estimate of scatter using Huber's weights

invC: the inverse of C

iter: nr of iterations

flag: flag (true/false) for convergence
}
\description{
computes M-estimator of scatter matrix for the n x p data matrix X
using the loss function 'Huber' or 't-loss' and for a given parameter of
the loss function (i.e., q for Huber's or degrees of freedom v for
the t-distribution).
}
\examples{

}
