% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_Auxi_order_wk.R
\name{order_wk}
\alias{order_wk}
\title{order_wk}
\usage{
order_wk(y)
}
\description{
order_wk
}
\note{
Location: .../Rrobuststp/R/SpectrumEstimation_Auxi_order_wk.r
}
\examples{

library(signal)
library(Rrobustsp)
library(zeallot)

y <- rnorm(5)

c(wr, PSD) <- order_wk(y)

}
