% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{rankflasso}
\alias{rankflasso}
\title{Computes the rank fused-Lasso regression estimates for given fused
penalty value lambda_2 and for a range of lambda_1 values}
\usage{
rankflasso(y, X, lambda1, lambda2, b0 = NULL, printitn = 0)
}
\arguments{
\item{y}{: numeric response N x 1 vector (real/complex)}

\item{X}{: numeric feature  N x p matrix (real/complex)}

\item{lambda1}{: positive penalty parameter for the Lasso penalty term}

\item{lambda2}{: positive penalty parameter for the fused Lasso penalty term}

\item{b0}{: numeric optional initial start (regression vector) of
iterations. If not given, we use LSE (when p>1).}

\item{printitn}{: print iteration number (default = 0, no printing)}
}
\value{
b      : numeric regression coefficient vector

iter   : positive integer, the number of iterations of IRWLS algorithm
}
\description{
Computes the rank fused-Lasso regression estimates for given fused
penalty value lambda_2 and for a range of lambda_1 values
}
\examples{

}
