% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{rankflassopath}
\alias{rankflassopath}
\title{rank fused Lasso regression}
\usage{
rankflassopath(y, X, lambda2, L = 120, eps = 0.001, printitn = F)
}
\arguments{
\item{y}{: numeric response N x 1 vector (real/complex)}

\item{X}{: numeric feature  N x p matrix (real/complex)}

\item{lambda2}{: positive penalty parameter for the fused Lasso penalty term}

\item{L}{: number of grid points for lambda1 (Lasso penalty)}

\item{eps}{: Positive scalar, the ratio of the smallest to the
largest Lambda value in the grid. Default is eps = 10^-4.}

\item{printitn}{: print iteration number (default = F, no printing)}
}
\value{
B: Fitted rank fused-Lasso regression coefficients, a p-by-(L+1) matrix,
          where p is the number of predictors (columns) in X, and L is
          the  number of Lambda values.

B0: estimates values of intercepts

lamgrid: = lambda parameters
}
\description{
Computes the rank fused-Lasso regression estimates for given fused
penalty value lambda_2 and for a range of lambda_1 values
}
\examples{

rankflassopath()
}
