% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_repeated_median_filter.R
\name{repeated_median_filter}
\alias{repeated_median_filter}
\title{Repeated Median filter}
\usage{
repeated_median_filter(x)
}
\arguments{
\item{x:}{numeric vector. The signal}
}
\value{
xFRM: Rpeated median filtered signal

ARM: Fourier coefficients for cosine

BRM: Fourier coefficients for sine
}
\description{
Implementation of
}
\details{
"High breakdown methods of time series analysis.
   Tatum, L.G., and Hurvich, C. M.
   Journal of the Royal Statistical Society. Series B (Methodological),
   pp. 881-896, 1993.
}
\note{
file in SpectrumEstimation_repeated_median_filter.R

Warning: Takes a long time
}
