% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{repmat}
\alias{repmat}
\title{repmat
 matlab like repmat, vector v is replicate s times columnwise into a matrix}
\usage{
repmat(v, s)
}
\description{
repmat
 matlab like repmat, vector v is replicate s times columnwise into a matrix
}
