% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_signcm.R
\name{signcm}
\alias{signcm}
\title{signcm}
\usage{
signcm(x, center = F)
}
\arguments{
\item{center}{: boolean. If True center data using spatial median
Default = False}

\item{X}{: N x p matrix}
}
\value{
C : Spatial covariance matrix

smed0 : spatial median. Computed only if center = True
}
\description{
calculates the spatial sign covariance matrix (:= SCM)
}
\examples{

x <- matrix(rnorm(9), 3, 3)

signcm(x)
}
