% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_spec_arma_est_bip.R
\name{spec_arma_est_bip_mm}
\alias{spec_arma_est_bip_mm}
\title{spec_arma_est_bip_mm}
\usage{
spec_arma_est_bip_mm(x, p, q, tolx = 5e-07)
}
\description{
spec_arma_est_bip_mm
}
\note{
File located in SpectrumEstimation_spec_arma_est_bip.R
}
