% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary.R
\name{tau_scale}
\alias{tau_scale}
\title{tau_scale}
\usage{
tau_scale(x)
}
\arguments{
\item{x}{}
}
\value{
scale
}
\description{
tau_scale
}
\note{
Location: .../Rrobustsp/R/dependentData_Auxiliary
}
