% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentData_Auxilliary_armaResid.R
\name{arma_tau_resid_sc}
\alias{arma_tau_resid_sc}
\title{arma_tau_resid_sc}
\usage{
arma_tau_resid_sc(x, beta_hat, p, q)
}
\description{
arma_tau_resid_sc
}
\note{
File is in dependentData_Auxiliary_armaResid.R
}
