% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{elemfits}
\alias{elemfits}
\title{[beta, w] = elemfits(x,y)
  elemfits compute the Nx(N-1)/2 elemental fits, i.e., intercepts b_{0,ij}
  and slopes b_{1,ij}, that define a line y = b_0+b_1 x that passes through
  the data points (x_i,y_i) and (x_j,y_j), i<j, where i, j in {1, ..., N}.
  and the respective weights | x_i - x_j |}
\usage{
elemfits(y, x)
}
\arguments{
\item{y}{: (numeric) N  vector of real-valued outputs (response vector)}

\item{x}{: (numeric) N  vector of inputs (feature vector)}
}
\value{
beta: (numeric) N*(N-1)/2 matrix of elemental fits

w: (numeric) N*(N-1)/2 matrix of weights
}
\description{
[beta, w] = elemfits(x,y)
  elemfits compute the Nx(N-1)/2 elemental fits, i.e., intercepts b_{0,ij}
  and slopes b_{1,ij}, that define a line y = b_0+b_1 x that passes through
  the data points (x_i,y_i) and (x_j,y_j), i<j, where i, j in {1, ..., N}.
  and the respective weights | x_i - x_j |
}
