% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_hubreg.R
\name{hubreg}
\alias{hubreg}
\title{hubreg}
\usage{
hubreg(y, X, c = NULL, sig0 = NULL, b0 = NULL, printitn = 0,
  iter_max = 2000, errortol = 1e-05)
}
\arguments{
\item{y:}{Numeric data vector of size N x 1 (output, respones)}

\item{X:}{Numeric data matrix of size N x p. Each row represents one
observation, and each column represents one predictor (feature).
If the model has an intercept, then first column needs to be a
vector of ones.}

\item{c:}{numeric threshold constant of Huber's function}

\item{sig0:}{(numeric) initial estimator of scale \cr
default = SQRT(1/(n-p)*RSS)}

\item{b0:}{initial estimator of regression (default: LSE)}

\item{printitn:}{print iteration number (default = 0, no printing)}

\item{iter_max:}{maximum number of iterations. \cr default = 2000}

\item{errortol:}{ERROR TOLERANCE FOR HALTING CRITERION. \cr default = 1e-5}
}
\value{
b1: the regression coefficient vector estimate

sig1: the estimate of scale

iter: the # of iterations
}
\description{
regression and scale using Huber's criterion
}
\details{
hubreg computes the joint M-estimates of regression and scale using
  Huber's criterion. Function works for both real- and complex-valued data.
}
\examples{

library(MASS)
y <- c(1.0347, 0.7269, -0.3034, 0.2939, -0.7873)
X <- matrix(c(0.884, -1.1471, -1.0689, -0.8095, -2.9443, 1.4384, 0.3252, -0.7549, 1.3703, -1.7115), 5, 2)

hubreg(y, X)
hubreg(y+1i, X)
hubreg(y+1i, X+1i)
hubreg(y, X+1i)

}
\references{
uses \code{\link[MASS]{ginv}} from the MASS package
}
