% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustFiltering_mParamEst.R
\name{m_param_est}
\alias{m_param_est}
\title{m_param_est}
\usage{
m_param_est(receive, C, parameter)
}
\arguments{
\item{recieve:}{is the received signal}

\item{Theta:}{contains the initial estimate}

\item{C:}{is the regression matrix of the model y = C*x + n}

\item{parameter.maxiters:}{maximal number of iterations}

\item{parameter.break:}{break condition}
}
\value{
th2: M-estimate of regression

Theta:  M-estimates of regression for all iterations

kk: iteration index

residual: residuals given Theta
}
\description{
The function computes an M-estimator of regression using the assymetric tanh score function.
}
\examples{

}
\references{
"Robust Statistics for Signal Processing"
  Zoubir, A.M. and Koivunen, V. and Ollila, E. and Muma, M.
  Cambridge University Press, 2018.

 "Robust Tracking and Geolocation for Wireless Networks in NLOS Environments."
  Hammes, U., Wolsztynski, E., and Zoubir, A.M.
  IEEE Journal on Selected Topics in Signal Processing, 3(5), 889-901, 2009.
}
