% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{madn}
\alias{madn}
\title{normalized median absolute deviation}
\usage{
madn(y)
}
\arguments{
\item{y:}{data vector of size N x 1}
}
\value{
sig: normalized median absolute deviations scale estimate
}
\description{
madn computes the normalized median absolute deviation estimate of
scale, i.e.,
}
\details{
mu_hat = arg min_mu SUM_i rho_TUK(y_i - mu)
}
