% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location_Scale.R
\name{Mloc}
\alias{Mloc}
\title{Mloc computes the M-estimates of location using an auxiliary scale
 estimate. It uses the iterative reweighted least squares (IRWLS) algorithm}
\usage{
Mloc(y, lossfun)
}
\arguments{
\item{y}{: (numeric) data vector of size N x 1}

\item{lossfun}{: (string) either 'huber' or 'tukey' to identify the desired
loss function}
}
\value{
mu_hat: M-estimate of location
}
\description{
Mloc computes the M-estimates of location using an auxiliary scale
 estimate. It uses the iterative reweighted least squares (IRWLS) algorithm
}
