% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location_Scale.R
\name{MlocTUK}
\alias{MlocTUK}
\title{Tukey's M-estimate of location}
\usage{
MlocTUK(y, c = 4.685, max_iters = 1000, tol_err = 1e-05)
}
\arguments{
\item{y}{: real valued data vector of size N x 1}

\item{c}{: tuning constant c>=0}
}
\value{
mu_hat : Tukey's M-estimate of location
}
\description{
Mloc_TUK computes Tukey's M-estimate of
  location, i.e.,
}
\details{
mu_hat = arg min_mu SUM_i rho_TUK(y_i - mu)
}
\note{
File location: Location_Scale.R
}
\examples{
MlocTUK(rnorm(5))

}
