% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{psihub}
\alias{psihub}
\title{Huber's score function psi: input is N x 1 data vector x which can be complex
or real and threshold contant c}
\usage{
psihub(x, c)
}
\description{
Huber's score function psi: input is N x 1 data vector x which can be complex
or real and threshold contant c
}
