% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{ranklasso}
\alias{ranklasso}
\title{ranklasso}
\usage{
ranklasso(y, X, lambda, b0 = NULL, printitn = F)
}
\arguments{
\item{y}{: numeric data vector of size N x 1 (output, respones)}

\item{X}{: numeric data matrix of size N x p (input, features)}

\item{lambda}{: penalty parameter (>= 0)}

\item{b0}{: numeric optional initial start (regression vector) of
iterations. If not given, we use LSE.}

\item{printitn}{: bool, print iteration number (default = F, no printing) and
other details}
}
\description{
ranklasso computes the rank (LAD-)regression estimate
}
\examples{

}
