% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{ranklassopath}
\alias{ranklassopath}
\title{ranklassopath}
\usage{
ranklassopath(y, X, L = 120, eps = 0.001, reltol = 1e-07,
  printitn = F)
}
\arguments{
\item{y:}{Numeric data vector of size N (output, respones)}

\item{X:}{Numeric data matrix of size N x p. Each row represents one
observation, and each column represents one predictor (feature).}

\item{L:}{Positive integer, the number of lambda values on the grid to be
used. The default is L=120.}

\item{eps:}{Positive scalar, the ratio of the smallest to the
largest Lambda value in the grid. Default is eps = 10^-3.}

\item{reltol:}{Convergence threshold for IRWLS. Terminate when successive
estimates differ in L2 norm by a rel. amount less than reltol.}

\item{printitn:}{print iteration number (default = F, no printing)}
}
\value{
B: Fitted RLAD-Lasso regression coefficients, a p-by-(L+1) matrix,
          where p is the number of predictors (columns) in X, and L is
          the  number of Lambda values.

B0: estimates values of intercepts

stats: structure with following fields:
            Lambda = lambda parameters in ascending order
            GMeAD = Mean Absolute Deviation (MeAD) of the residuals
            gBIC = generalized Bayesian information criterion (gBIC) value
                 for each lambda parameter on the grid.
}
\description{
ranklassopath computes the rank LAD-Lasso regularization path (over grid
                                                                 of penalty parameter values). Uses IRWLS algorithm.
}
\examples{
y <- 1:5
X <- matrix(rnorm(15),5,3)

tmp <- ranklassopath(y, X)


tmp[[1]] # B
tmp[[2]] # B0
tmp[[3]]['GMeAD'] # stats fiedl GMeAD


}
