% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{rladreg}
\alias{rladreg}
\title{rladreg}
\usage{
rladreg(y, X, b0 = NULL, printitn = F)
}
\arguments{
\item{y:}{numeric response N x 1 vector (real/complex)}

\item{X:}{numeric feature  N x p matrix (real/complex)}

\item{b0:}{numeric optional initial start of the regression vector for
IRWLS algorithm. If not given, we use LSE (when p>1).}

\item{printitn:}{print iteration number (default = 0, no printing) and
other details}
}
\value{
b1: numeric the regression coefficient vector

iter: (numeric) # of iterations (given when IRWLS algorithm is used)
}
\description{
computes the LAD regression estimate
}
