% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumEstimation_spec_arma_est_bip.R
\name{spec_arma_est_bip_s}
\alias{spec_arma_est_bip_s}
\title{spec_arma_est_bip_s}
\usage{
spec_arma_est_bip_s(x, p, q)
}
\description{
spec_arma_est_bip_s
}
\note{
File located in SpectrumEstimation_spec_arma_est_bip.R
}
